Game Development Research Project
I'm working on a project which is used to design video game levels by
using Deep Learning , Machine Learning and Artificial Intelligence.The
tools used for this project are Caffe , Theano. We are doing this
research in order to introduce a new course called Deep Learning for the
Spring 2017 Semester at the USC Viterbi School of Engineering.